voice-assistant/requirements.txt
numpy>=1.26
sounddevice>=0.5
faster-whisper>=1.1
google-genai>=1.0
pytest>=8.0
